Python_Learnibg_asssistant/pyproject.toml
[project]
name = "python-learnibg-asssistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.8.0",
    "openai-agents>=0.3.0",
]
